requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
torch
python-multipart
streamlit


ultralytics>=8.1.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

streamlit>=1.30.0
streamlit-folium>=0.18.0
folium>=0.15.0

google-genai>=0.3.0
google-adk>=0.1.0

requests>=2.31.0

# ==============================
# Utilities
# ==============================
python-dotenv>=1.0.0
loguru>=0.7.2
